Fix various wording problems with template-instantiation article (#97)
A few things in the article aren't quite right and should be corrected:

1. A template is an
[entity](https://eel.is/c++draft/basic.pre#def:entity) by definition.
However, it would be correct to say that a function template is not a
function.

2. To my knowledge, "deriving to" or "deriving into" isn't correct
English. However, you can say "turning into", which most would consider
simpler language anyway.

3. It's called "function template", but the heading says "template
function". We don't need to be pedantically correct in these articles,
but at least we should use the terminology from the
standard/cppreference.